fix: complete norm exec.c and complete norm split function fixes: #67
