+++ src/exec/argv_utils.c
@@ -1,0 +1,64 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   argv_utils.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yshimoda <yshimoda@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/05/07 11:37:07 by yshimoda          #+#    #+#             */
+/*   Updated: 2023/05/07 11:37:11 by yshimoda         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+#include <stdlib.h>
+#include <string.h>
+
+size_t	argv_len(t_tk *token)
+{
+	size_t	len;
+
+	len = 0;
+	while (token)
+	{
+		len++;
+		token = token->next;
+	}
+	return (len);
+}
+
+char	**make_argv(t_tk *token)
+{
+	size_t	i;
+	size_t	len;
+	char	**argv;
+
+	if (!token)
+		return (NULL);
+	i = 0;
+	len = argv_len(token);
+	argv = calloc(len + 1, sizeof(char *));
+	if (!argv)
+		assert_error("malloc\n");
+	while (token->kind != TK_EOF)
+	{
+		argv[i] = strdup(token->word);
+		if (!argv[i])
+			assert_error("strdup\n");
+		token = token->next;
+		i++;
+	}
+	return (argv);
+}
+
+void	free_argv(char **argv)
+{
+	size_t	i;
+
+	if (!argv)
+		return ;
+	i = 0;
+	while (argv[i])
+		free(argv[i++]);
+	free(argv);
+}
