+++ src/exec/libft.c
@@ -1,0 +1,101 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   libft.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yshimoda <yshimoda@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/05/07 11:38:12 by yshimoda          #+#    #+#             */
+/*   Updated: 2023/05/07 11:38:14 by yshimoda         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+#include <stdlib.h>
+#include <string.h>
+
+static size_t	get_size(int n)
+{
+	size_t	len;
+
+	len = 0;
+	if (!n)
+		return (1);
+	if (n < 0)
+		len++;
+	while (n)
+	{
+		n /= 10;
+		len++;
+	}
+	return (len);
+}
+
+char	*itoa(int n)
+{
+	char	*s;
+	size_t	len;
+	int		num;
+
+	len = get_size(n);
+	s = malloc(sizeof(char) * (len + 1));
+	if (!s)
+		return (NULL);
+	s[len] = '\0';
+	num = n;
+	while (len--)
+	{
+		if (num < 0)
+			s[len] = -(num % 10) + '0';
+		else
+			s[len] = num % 10 + '0';
+		num /= 10;
+	}
+	if (n < 0)
+		*s = '-';
+	return (s);
+}
+
+char	*strjoin_three(char const *s1, char const *s2, char const *s3)
+{
+	char	*ptr;
+	char	*save;
+
+	if (!s1 || !s2 || !s3)
+		return (NULL);
+	ptr = (char *)malloc(strlen(s1) + strlen(s2) + strlen(s3) + 1);
+	if (!ptr)
+		return (NULL);
+	save = ptr;
+	while (*s1)
+		*ptr++ = *s1++;
+	while (*s2)
+		*ptr++ = *s2++;
+	while (*s3)
+		*ptr++ = *s3++;
+	*ptr = 0;
+	return (save);
+}
+
+char	*strjoin(char const *s1, char const *s2)
+{
+	char	*ptr;
+	char	*save;
+
+	if (!s1 && !s2)
+		return (NULL);
+	if (!s1)
+		return (strdup(s2));
+	if (!s2)
+		return (strdup(s1));
+	ptr = (char *)malloc(strlen(s1) + strlen(s2) + 1);
+	if (!ptr)
+		return (NULL);
+	save = ptr;
+	while (*s1)
+		*ptr++ = *s1++;
+	while (*s2)
+		*ptr++ = *s2++;
+	*ptr = 0;
+	return (save);
+}
